# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Worldwide Rapid Task For Money - Verification Automation Engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/transactions:
    get:
      operationId: listTransactions
      tags: [wallet]
      summary: List wallet transactions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /wallet/deposit:
    post:
      operationId: depositFunds
      tags: [wallet]
      summary: Manual deposit funds
      description: Disabled by default. Use /payments/crypto/deposit for production client funding.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "200":
          description: Deposit initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "410":
          description: Manual deposits disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /wallet/withdraw:
    post:
      operationId: withdrawFunds
      tags: [wallet]
      summary: Withdraw funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "200":
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # Dashboard summaries (wow endpoints)
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Transaction:
      type: object
      required: [id, walletId, type, amount, status, createdAt]
      properties:
        id:
          type: integer
        walletId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, task_reward, task_payment, refund, bonus]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed, cancelled]
        reference:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TransactionList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DepositInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 1
        currency:
          type: string
        method:
          type: string
...
    WithdrawInput:
      type: object
      required: [amount, method]
      properties:
        amount:
          type: number
          minimum: 1
        method:
          type: string
        accountDetails:
          type: string

    # Dashboard summaries
...